[settings] Adds recipe
Use this recipe to include application based settings. Add the elephant::settings recipe at the end of the run list to make sure all recipes are loaded.
Using mac_os_x::settings has the side effect of always trying to import all settings, including the ones for apps that were not in the recipes run list.
A better practice would be to add something like include_at_end_of_runlist('elephant::settings') in these recipes, but afaik this is not available in chef.